修改代码。

+++ yn-job/src/main/java/com/yn/dao/TaskExecuteRecordMapper.java
@@ -16,7 +16,7 @@ public interface TaskExecuteRecordMapper {
 
     int deleteByPrimaryKey(Long id);
 
-    int insert(TaskExecuteRecord record);
+    Long insert(TaskExecuteRecord record);
 
     int insertSelective(TaskExecuteRecord record);
